+++ src/lib/utils/jsx.ts
@@ -18,16 +18,16 @@ import type {
     JsxElement,
     JsxChildren,
     JsxComponent,
-} from "./jsx.elements.js";
-import { JsxFragment as Fragment } from "./jsx.elements.js";
+} from "./jsx.elements";
+import { JsxFragment as Fragment } from "./jsx.elements";
 
 // Backwards compatibility until 0.24
 export type {
     JsxElement as Element,
     JsxChildren as Children,
     JsxComponent,
-} from "./jsx.elements.js";
-export { JsxFragment as Fragment } from "./jsx.elements.js";
+} from "./jsx.elements";
+export { JsxFragment as Fragment } from "./jsx.elements";
 
 /**
  * Used to inject HTML directly into the document.