+++ src/test/converter2/issues/gh1875.ts
@@ -1,0 +1,3 @@
+export declare function test(this: typeof globalThis, param: string): string;
+
+export declare function test2(this, param: string): string;