+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 -   If an exported symbol has multiple declarations, TypeDoc will now check all appropriate declarations for comments, and warn if more than one declaration contains a comment, #1855.
 -   Improved support for JSDoc style `@example` tags. If the tag content does not include a code block, TypeDoc now follows VSCode's behavior of treating the entire block as a code block, #135.
 -   TypeDoc will now render members marked with `@deprecated` with a line through their name, #1381.
+-   TypeDoc will now warn if part of a comment will be overwritten due to use of `@inheritDoc` instead of silently dropping part of the comment
 
 ### Bug Fixes