+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Features
 
+-   Add support for TypeScript 4.6, #1877.
 -   Support copying `@param` comments for nested members that target union and intersection types, #1876.
 
 ## v0.22.12 (2022-02-20)